Factor out get source tests (#624)
